Start switching over to Events.
Instead of relying on year to link leads and showcase (and others) to a specific
BFIG instance, allow users to create events which they will be part of. This has
replaced year everywhere, and now year columns need to be removed from the database.

I also started working on the benefits manager. I've designed the schema as well
as the defaults, and the list page.